[9.x] WFLY-4601 Upgrade Infinispan to 7.2.1.Final

+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/CacheAddHandler.java
@@ -260,7 +260,7 @@ public class CacheAddHandler extends AbstractAddStepHandler {
                 builder.eviction().strategy(strategy);
 
                 if (strategy.isEnabled()) {
-                    final int maxEntries = EvictionResourceDefinition.MAX_ENTRIES.resolveModelAttribute(context, eviction).asInt();
+                    final long maxEntries = EvictionResourceDefinition.MAX_ENTRIES.resolveModelAttribute(context, eviction).asLong();
                     builder.eviction().maxEntries(maxEntries);
                 }
             }
